# LoggingInvocationEventHandler precomputes level information