--- a/changelog/@unreleased/pr-997.v2.yml
+++ b/changelog/@unreleased/pr-997.v2.yml
@@ -1,0 +1,5 @@
+type: improvement
+improvement:
+  description: LoggingInvocationEventHandler precomputes level information
+  links:
+  - https://github.com/palantir/tritium/pull/997