--- a/tritium-lib/src/test/java/com/palantir/tritium/event/log/LoggingInstrumentationTest.java
+++ b/tritium-lib/src/test/java/com/palantir/tritium/event/log/LoggingInstrumentationTest.java
@@ -105,7 +105,7 @@ public class LoggingInstrumentationTest {
                 .withHandler(handler)
                 .build();
 
-        assertThat(LoggingInvocationEventHandler.isEnabled(logger, level)).isTrue();
+        assertThat(handler.isEnabled()).isTrue();
 
         assertThat(delegate.invocationCount()).isZero();