--- a/tritium-slf4j/src/test/java/com/palantir/tritium/event/log/LoggingInvocationEventHandlerTest.java
+++ b/tritium-slf4j/src/test/java/com/palantir/tritium/event/log/LoggingInvocationEventHandlerTest.java
@@ -94,13 +94,6 @@ public class LoggingInvocationEventHandlerTest {
 
     @Test
     @SuppressWarnings("NullAway") // explicitly testing null handling
-    public void testNullIsEnabled() {
-        assertThatExceptionOfType(NullPointerException.class)
-                .isThrownBy(() -> LoggingInvocationEventHandler.isEnabled(getLogger(), null));
-    }
-
-    @Test
-    @SuppressWarnings("NullAway") // explicitly testing null handling
     public void testNullLevel() {
         assertThatExceptionOfType(NullPointerException.class)
                 .isThrownBy(() -> new LoggingInvocationEventHandler(getLogger(), null));