@carterkozak 2021-02-22
Just pushing this since I toyed with it a little in my free time last week. No rush to review, opening a PR so I don't forget about it.