@carterkozak 2021-02-23
### Before
logging enabled to /dev/null
```
Benchmark                                                 (mode)  Mode  Cnt     Score     Error  Units
ProxyBenchmark.instrumentedWithPerformanceLogging     BYTE_BUDDY  avgt    5  2353.263 ±  40.420  ns/op
ProxyBenchmark.instrumentedWithPerformanceLogging  DYNAMIC_PROXY  avgt    5  2304.355 ± 194.096  ns/op
```
logging disabled (via level):
```
Benchmark                                                 (mode)  Mode  Cnt   Score   Error  Units
ProxyBenchmark.instrumentedWithPerformanceLogging     BYTE_BUDDY  avgt    5   4.884 ± 0.047  ns/op
ProxyBenchmark.instrumentedWithPerformanceLogging  DYNAMIC_PROXY  avgt    5  11.419 ± 0.847  ns/op
```

### After
logging enabled to /dev/null
```
Benchmark                                                 (mode)  Mode  Cnt     Score     Error  Units
ProxyBenchmark.instrumentedWithPerformanceLogging     BYTE_BUDDY  avgt    5  2301.592 ±  54.826  ns/op
ProxyBenchmark.instrumentedWithPerformanceLogging  DYNAMIC_PROXY  avgt    5  2351.683 ± 189.756  ns/op
```
logging disabled (via level):
```
Benchmark                                                 (mode)  Mode  Cnt   Score   Error  Units
ProxyBenchmark.instrumentedWithPerformanceLogging     BYTE_BUDDY  avgt    5   4.717 ± 0.322  ns/op
ProxyBenchmark.instrumentedWithPerformanceLogging  DYNAMIC_PROXY  avgt    5  10.931 ± 0.956  ns/op
```

The initial set of results with logging enabled is dominated by slf4j-simple, and results for the expected path (log level is not enabled) both sets of results appear a bit faster.